requirements.txt
numpy
pydicom
scipy
Pillow